Refactor(#16) : Payment 엔터티 칼럼 리펙토링
